requirement.txt
absl-py==2.1.0
aiohttp==3.9.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.12.3
bibtexparser==1.4.1
cachetools==5.3.3
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
contourpy==1.1.1
cycler==0.12.1
filelock==3.15.4
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.6.1
gdown==5.2.0
google-auth==2.32.0
google-auth-oauthlib==0.4.6
grpcio==1.64.1
huggingface-hub==0.23.4
idna==3.7
importlib-resources==6.4.0
jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.5
lightning==2.3.3
lightning-utilities==0.11.3.post0
markdown==3.6
markupsafe==2.1.5
matplotlib==3.7.5
mpmath==1.3.0
multidict==6.0.5
networkx==3.1
nose==1.3.7
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.82
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
packaging==24.1
pandas==2.0.3
pdf2image==1.17.0
pillow==10.4.0
protobuf==5.27.2
pyasn1==0.6.0
pyasn1-modules==0.4.0
pycparser==2.22
pyparsing==3.1.2
pysocks==1.7.1
python-dateutil==2.9.0.post0
pytorch-lightning==2.3.3
pytorch-model-summary==0.1.2
pytz==2024.1
pyyaml==6.0.1
regex==2024.5.15
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
safetensors==0.4.3
scikit-learn==1.3.2
scipy==1.10.1
six==1.16.0
soundfile==0.12.1
soupsieve==2.5
sympy==1.13.0
tensorboard==2.12.0
tensorboard-data-server==0.7.2
tensorboard-plugin-wit==1.8.1
tensorboardx==2.6.2.2
threadpoolctl==3.5.0
timm==1.0.7
tokenizers==0.19.1
torch==2.3.1
torchaudio==2.3.1
torchmetrics==1.4.0.post0
torchvision==0.18.1
tqdm==4.66.4
transformers==4.42.3
triton==2.3.1
typing-extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
werkzeug==3.0.3
yarl==1.9.4
zipp==3.19.2
